move static playlist calc logic to SpotifyPlaylistCalculator
